商品接口提交

+++ trade-center/trade-center-servie/src/main/java/com/ai/bdex/dataexchange/tradecenter/service/interfaces/gds/IGdsCatSV.java
@@ -6,5 +6,5 @@ import com.ai.bdex.dataexchange.tradecenter.dao.model.GdsCat;
  * Created by yx on 2017/4/17.
  */
 public interface IGdsCatSV {
-    public GdsCat queryGdsCatById(Long catId) throws Exception;
+    public GdsCat queryGdsCatById(Integer catId) throws Exception;
 }
